Implement reduce function
- Implemented reduce function and unit tests for it.
- Reduce function takes 3 arguments (iterable, reducer and
initialValue). Converts iterable to an array if it's not.
- Throws an TypeError if the reducer argument is not a function.
- If the initialValue is set, it is initially stored in the accumulator,
which stores the value of the execution of the function passed in the
reducer argument. If initialValue is undefined, the first value in the
iterable array is store in the accumulator.
- Then, for each value in the array, the reducer function is executed,
it takes 3 arguments (an accumulator, the currentValue in the array and
it's currentIndex);
- After the for loop, the accumulator is returned.